llm2fx-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
jinja2==3.1.3
librosa==0.11.0
soundfile==0.13.1
numpy>=1.20.0
scipy>=1.7.0
